Move users' create handler to separate function
